Split synchronization and stacking modules
