+++ scinoephile/cli/scinoephile_cli.py
@@ -32,9 +32,6 @@ SCINOEPHILE_LOCALIZATIONS: dict[str, dict[str, str]] = {
         "another": "计算一个序列相对于另一个序列的字符错误率（CER）",
         "calculate the diff between two series": "计算两个序列之间的差异",
         "command-line interface for Scinoephile": "Scinoephile 命令行界面",
-        "combine two series into the top and bottom of a synchronized series": (
-            "将两个序列合并为上下行同步字幕"
-        ),
         "fuse OCR output for a selected language": "融合所选语言的 OCR 输出",
         "inspect and extract media streams": "检查并提取媒体流",
         "list all subcommands and exit": "列出所有子命令并退出",
@@ -51,6 +48,9 @@ SCINOEPHILE_LOCALIZATIONS: dict[str, dict[str, str]] = {
         "shift and stretch the timings of one subtitle series to match another": (
             "平移并拉伸一个字幕序列的时间轴以匹配另一个序列"
         ),
+        "stack two series into top and bottom subtitle lines": (
+            "将两个序列堆叠为上下行字幕"
+        ),
         "validate OCR text against subtitle images": "对照字幕图像校验 OCR 文本",
     },
     "zh-hant": {
@@ -61,9 +61,6 @@ SCINOEPHILE_LOCALIZATIONS: dict[str, dict[str, str]] = {
         "another": "計算一個序列相對於另一個序列的字元錯誤率（CER）",
         "calculate the diff between two series": "計算兩個序列之間的差異",
         "command-line interface for Scinoephile": "Scinoephile 命令列介面",
-        "combine two series into the top and bottom of a synchronized series": (
-            "將兩個序列合併為上下行同步字幕"
-        ),
         "fuse OCR output for a selected language": "融合所選語言的 OCR 輸出",
         "inspect and extract media streams": "檢查並提取媒體流",
         "list all subcommands and exit": "列出所有子命令並結束",
@@ -79,6 +76,9 @@ SCINOEPHILE_LOCALIZATIONS: dict[str, dict[str, str]] = {
         ): "Scinoephile 是用於處理中文、英文與雙語字幕的應用程式。",
         "shift and stretch the timings of one subtitle series to match another": (
             "平移並拉伸一個字幕序列的時間軸以匹配另一個序列"
+        ),
+        "stack two series into top and bottom subtitle lines": (
+            "將兩個序列堆疊為上下行字幕"
         ),
         "validate OCR text against subtitle images": "對照字幕影像驗證 OCR 文字",
     },
