--- scinoephile/core/synchronization.py
@@ -1,779 +1,0 @@
-#  Copyright 2017-2026 Karl T Debiec. All rights reserved. This software may be modified
-#  and distributed under the terms of the BSD license. See the LICENSE file for details.
-"""Core code related to synchronization of subtitles."""
-
-from __future__ import annotations
-
-from dataclasses import dataclass
-from enum import StrEnum
-from logging import getLogger
-from pprint import pformat
-from statistics import mean, median, pstdev
-
-import numpy as np
-
-from .exceptions import ScinoephileError
-from .pairs import get_block_pairs_by_pause, get_pair_strings
-from .subtitles import Series, Subtitle, get_concatenated_series
-
-__all__ = [
-    "SyncGroup",
-    "SyncOffsetSample",
-    "SyncOffsetStats",
-    "SyncTimingMode",
-    "are_series_one_to_one",
-    "get_overlap_string",
-    "get_sync_groups",
-    "get_sync_groups_string",
-    "get_sync_offset_stats",
-    "get_sync_overlap_matrix",
-    "get_synced_series",
-    "get_synced_series_from_groups",
-]
-
-logger = getLogger(__name__)
-
-SyncGroup = tuple[list[int], list[int]]
-"""Group of subtitles; items are indexes in first and second series, respectively."""
-
-
-@dataclass(frozen=True)
-class SyncOffsetSample:
-    """One paired sync group used for subtitle offset estimation."""
-
-    block_index: int
-    """Zero-based block index containing the paired sync group."""
-
-    group_index: int
-    """Zero-based sync group index within the block."""
-
-    anchor_indexes: tuple[int, ...]
-    """Block-local subtitle indexes from the anchor series."""
-
-    mobile_indexes: tuple[int, ...]
-    """Block-local subtitle indexes from the mobile series."""
-
-    offset_ms: float
-    """Mobile group midpoint minus anchor group midpoint, in milliseconds."""
-
-
-@dataclass(frozen=True)
-class SyncOffsetStats:
-    """Statistics for subtitle offset estimation."""
-
-    sample_count: int
-    """Number of paired sync groups included in the estimate."""
-
-    skipped_group_count: int
-    """Number of unpaired sync groups skipped during estimation."""
-
-    mean_ms: float
-    """Mean mobile-minus-anchor offset, in milliseconds."""
-
-    median_ms: float
-    """Median mobile-minus-anchor offset, in milliseconds."""
-
-    stdev_ms: float
-    """Population standard deviation of offsets, in milliseconds."""
-
-    min_ms: float
-    """Minimum mobile-minus-anchor offset, in milliseconds."""
-
-    max_ms: float
-    """Maximum mobile-minus-anchor offset, in milliseconds."""
-
-    samples: tuple[SyncOffsetSample, ...]
-    """Offset samples included in the estimate."""
-
-
-class SyncTimingMode(StrEnum):
-    """Timing modes for synchronized subtitles."""
-
-    TOP = "top"
-    """Use top subtitle timing where available."""
-    BOTTOM = "bottom"
-    """Use bottom subtitle timing where available."""
-    OUTER = "outer"
-    """Use the full outer timing span of paired subtitles."""
-
-
-def are_series_one_to_one(one: Series, two: Series) -> bool:
-    """Check whether two series are one-to-one matched.
-
-    This is useful for preparing test cases, specifically for excluding one-to-one
-    mappings, which are simple and not of interest for testing.
-
-    Arguments:
-        one: First series to compare
-        two: Second series to compare
-    Returns:
-        Whether all subtitles are one-to-one matches between the two series
-    """
-    if len(one) != len(two):
-        return False
-
-    for one_sub, two_sub in zip(one, two):
-        if one_sub.start != two_sub.start or one_sub.end != two_sub.end:
-            return False
-
-    return True
-
-
-def get_overlap_string(overlap: np.ndarray) -> str:
-    """Get string representation of overlap matrix between two series.
-
-    1-indexed to match SRT.
-
-    Arguments:
-        overlap: Overlap matrix
-    Returns:
-        String representation of overlap matrix
-    """
-    max_items = 1_000_000
-    matrix = np.array2string(
-        overlap,
-        precision=2,
-        suppress_small=True,
-        max_line_width=max_items,
-        threshold=max_items,
-        edgeitems=max_items,
-    )
-    matrix = matrix.replace("0.  ", "____").replace("[", " ").replace("]", " ")
-    columns = [f"{j:>5}" for j in range(1, overlap.shape[1] + 1)]
-    lines = ["  " + "".join(columns)]
-    for i, row in enumerate(matrix.split("\n")):
-        lines += [f"{i + 1:>2}" + row]
-    return "\n".join(lines)
-
-
-def get_sync_groups_string(sync_groups: list[SyncGroup]) -> str:
-    """Get a string representation of sync groups.
-
-    Arguments:
-        sync_groups: Sync groups to represent
-    Returns:
-        String representation of the sync groups
-    """
-    one_indexed_sync_groups = []
-    for sync_group in sync_groups:
-        one_indexed_group = (
-            [i + 1 for i in sync_group[0]],
-            [j + 1 for j in sync_group[1]],
-        )
-        one_indexed_sync_groups.append(one_indexed_group)
-    return pformat(one_indexed_sync_groups, width=120, indent=2)
-
-
-def get_sync_groups(one: Series, two: Series, cutoff: float = 0.16) -> list[SyncGroup]:
-    """Distribute subtitles from two series into sync groups based on their overlap.
-
-    Subtitles are grouped based on the overlap between them. Subtitles are treated as
-    Gaussians whose centers are the midpoints of their start and end times and whose
-    standard deviations are one quarter of their duration. The overlaps between all
-    subtitles in the two series are calculated and stored in a matrix whose rows are
-    subtitles in the first series and whose columns are subtitles in the second series.
-    The matrix is then iteratively pruned by zeroing cells below an increasing cutoff
-    value. Within each row, each value is divided by the max value in that row, and if
-    the result is less than the cutoff, the cell is zeroed. The same process is then
-    repeated for columns. The process repeats with an increasing cutoff until each
-    non-zero value in the matrix is either the only value in its row, the only value in
-    its column, or both, which provides a clean assignment of subtitles to sync groups.
-
-    The cutoff starts at a minimum value, which defaults to 0.16. This serves as a lower
-    bound; if a subtitle in series two overlaps less than this amount with a subtitle
-    in series one, it is included without a partner.
-
-    Arguments:
-        one: First series
-        two: Second series
-        cutoff: Initial overlap cutoff used to adjust overlap matrix
-    Returns:
-        List of sync groups, each of which is a list of two lists of subtitle indexes,
-        the first list corresponding to subtitles in series one and the second list
-        corresponding to subtitles in series two.
-    """
-    if len(one) == 0:
-        return []
-    if len(two) == 0:
-        return [([i], []) for i in range(len(one))]
-
-    overlap = get_sync_overlap_matrix(one, two)
-    logger.debug(f"OVERLAP:\n{get_overlap_string(overlap)}")
-
-    max_cutoff = 1.0
-    sync_groups = None
-    while True:
-        try:
-            sync_groups = _get_sync_groups(one, two, overlap.copy(), cutoff)
-        except ScinoephileError:
-            cutoff += 0.01
-            if cutoff > max_cutoff:
-                raise ScinoephileError(
-                    f"Failed to compute sync groups: cutoff exceeded {max_cutoff}. "
-                    f"Final cutoff: {cutoff:.2f}. "
-                    f"Overlap matrix shape: {overlap.shape}. "
-                    f"This may indicate malformed or incompatible subtitle timing."
-                )
-            continue
-        break
-
-    return sync_groups
-
-
-def get_sync_offset_stats(
-    anchor: Series,
-    mobile: Series,
-    sync_cutoff: float = 0.16,
-    pause_length: int = 3000,
-) -> SyncOffsetStats:
-    """Estimate subtitle offset statistics between already-close series.
-
-    The offset is computed as mobile timing minus anchor timing. Positive values mean
-    the mobile series is later than the anchor series and should be shifted earlier to
-    align with it.
-
-    Arguments:
-        anchor: series used as timing reference
-        mobile: series to compare against the anchor
-        sync_cutoff: initial overlap cutoff used to form sync groups
-        pause_length: pause length in milliseconds used to split subtitle blocks
-    Returns:
-        offset statistics computed from paired sync groups
-    Raises:
-        ScinoephileError: if no paired sync groups are available
-    """
-    samples: list[SyncOffsetSample] = []
-    skipped_group_count = 0
-    block_pairs = get_block_pairs_by_pause(anchor, mobile, pause_length=pause_length)
-
-    for block_index, (anchor_block, mobile_block) in enumerate(block_pairs):
-        if not anchor_block.events or not mobile_block.events:
-            skipped_group_count += len(anchor_block.events) + len(mobile_block.events)
-            continue
-
-        groups = get_sync_groups(anchor_block, mobile_block, cutoff=sync_cutoff)
-        for group_index, (anchor_indexes, mobile_indexes) in enumerate(groups):
-            if not anchor_indexes or not mobile_indexes:
-                skipped_group_count += 1
-                continue
-
-            anchor_subs = [anchor_block.events[i] for i in anchor_indexes]
-            mobile_subs = [mobile_block.events[i] for i in mobile_indexes]
-            offset_ms = _get_subtitle_group_midpoint(
-                mobile_subs
-            ) - _get_subtitle_group_midpoint(anchor_subs)
-            samples.append(
-                SyncOffsetSample(
-                    block_index=block_index,
-                    group_index=group_index,
-                    anchor_indexes=tuple(anchor_indexes),
-                    mobile_indexes=tuple(mobile_indexes),
-                    offset_ms=offset_ms,
-                )
-            )
-
-    if not samples:
-        raise ScinoephileError(
-            "No paired sync groups found; cannot estimate subtitle offset"
-        )
-
-    offsets = [sample.offset_ms for sample in samples]
-    return SyncOffsetStats(
-        sample_count=len(samples),
-        skipped_group_count=skipped_group_count,
-        mean_ms=mean(offsets),
-        median_ms=median(offsets),
-        stdev_ms=pstdev(offsets),
-        min_ms=min(offsets),
-        max_ms=max(offsets),
-        samples=tuple(samples),
-    )
-
-
-def get_sync_overlap_matrix(one: Series, two: Series) -> np.ndarray:
-    """Quantify the overlap between two series and compile the results in a matrix.
-
-    Arguments:
-        one: First series
-        two: Second series
-    Returns:
-        Two-dimensional array whose rows correspond to subtitle indexes within series
-        one, whose columns correspond to subtitle indexes within series two, and whose
-        values are the proportion of each subtitle in series two which overlaps with
-        each subtitle in series one.
-    Raises:
-        ScinoephileError: if any subtitle has zero or negative duration
-    """
-    # Validate subtitle durations in series one
-    for i, event in enumerate(one.events):
-        duration = event.end - event.start
-        if duration <= 0:
-            raise ScinoephileError(
-                f"Subtitle {i + 1} in series one has invalid duration "
-                f"(start={event.start}, end={event.end}, duration={duration}). "
-                f"Subtitles must have positive duration for synchronization."
-            )
-
-    # Validate subtitle durations in series two
-    for i, event in enumerate(two.events):
-        duration = event.end - event.start
-        if duration <= 0:
-            raise ScinoephileError(
-                f"Subtitle {i + 1} in series two has invalid duration "
-                f"(start={event.start}, end={event.end}, duration={duration}). "
-                f"Subtitles must have positive duration for synchronization."
-            )
-
-    one_mu = np.array([e.start + (e.end - e.start) / 2 for e in one])
-    one_sigma = np.array([(e.end - e.start) / 4 for e in one])
-    two_mu = np.array([e.start + (e.end - e.start) / 2 for e in two])
-    two_sigma = np.array([(e.end - e.start) / 4 for e in two])
-
-    mu_diff_sq = (one_mu[:, np.newaxis] - two_mu[np.newaxis, :]) ** 2
-    sigma_sq_sum = one_sigma[:, np.newaxis] ** 2 + two_sigma[np.newaxis, :] ** 2
-    overlap = np.exp(-mu_diff_sq / (2 * sigma_sq_sum))
-
-    return overlap
-
-
-def get_synced_series(
-    one: Series,
-    two: Series,
-    sync_cutoff: float = 0.16,
-    pause_length: int = 3000,
-    timing_mode: SyncTimingMode = SyncTimingMode.OUTER,
-) -> Series:
-    """Compile synchonized subtitles from two series.
-
-    Arguments:
-        one: First Series
-        two: Second Series
-        sync_cutoff: Initial overlap cutoff used to form sync groups
-        pause_length: Pause length in milliseconds used to split subtitle blocks
-        timing_mode: Timing mode used for synchronized subtitles
-    Returns:
-        Synchonized subtitles
-    """
-    synced_blocks = []
-
-    logger.info(
-        f"Synchronizing series with sync_cutoff={sync_cutoff} "
-        f"and pause_length={pause_length}"
-    )
-
-    block_pairs = get_block_pairs_by_pause(one, two, pause_length=pause_length)
-    for one_block, two_block in block_pairs:
-        one_str, two_str = get_pair_strings(one_block, two_block)
-        logger.debug(f"ONE:\n{one_str}")
-        logger.debug(f"TWO:\n{two_str}")
-
-        groups = get_sync_groups(one_block, two_block, cutoff=sync_cutoff)
-        logger.info(f"SYNC GROUPS:\n{pformat(groups, width=1000)}")
-
-        synced_block = get_synced_series_from_groups(
-            one_block, two_block, groups, timing_mode=timing_mode
-        )
-        logger.info(f"SYNCED SUBTITLES:\n{synced_block.to_simple_string()}")
-        synced_blocks.append(synced_block)
-
-    synced = get_concatenated_series(synced_blocks)
-    return _get_series_without_overlaps(synced)
-
-
-def get_synced_series_from_groups(
-    one: Series,
-    two: Series,
-    groups: list[SyncGroup],
-    timing_mode: SyncTimingMode = SyncTimingMode.OUTER,
-) -> Series:
-    """Compile synchronized subtitles from two series based on sync groups.
-
-    Arguments:
-        one: First series
-        two: Second series
-        groups: Sync groups including the indexes of subtitles in each series
-        timing_mode: Timing mode used for synchronized subtitles
-    Returns:
-        Series whose subtitles are composed of the text of the subtitles from the two
-        input series as indicated by the sync groups
-    """
-    synced = Series()
-
-    for group in groups:
-        one_subs = [one.events[i] for i in group[0]]
-        two_subs = [two.events[i] for i in group[1]]
-
-        # Zero to one mapping
-        if len(one_subs) == 0 and len(two_subs) == 1:
-            synced.events.append(two_subs[0])
-            continue
-
-        # One to zero mapping
-        if len(one_subs) == 1 and len(two_subs) == 0:
-            synced.events.append(one_subs[0])
-            continue
-
-        # One to one mapping
-        if len(one_subs) == 1 and len(two_subs) == 1:
-            one_text = one_subs[0].text
-            two_text = two_subs[0].text
-            start, end = _get_sync_group_timings(one_subs, two_subs, 1, timing_mode)[0]
-            synced.events.append(
-                Subtitle(
-                    start=start,
-                    end=end,
-                    text=(
-                        f"{one_text}\\N{two_text}"
-                        if one_text and two_text
-                        else one_text or two_text
-                    ),
-                )
-            )
-            continue
-
-        # Many to one mapping
-        if len(one_subs) > 1 and len(two_subs) == 1:
-            two_sub = two_subs[0]
-            timings = _get_sync_group_timings(
-                one_subs, two_subs, len(one_subs), timing_mode
-            )
-
-            for one_sub, (start, end) in zip(one_subs, timings):
-                one_text = one_sub.text
-                two_text = two_sub.text
-                synced.events.append(
-                    Subtitle(
-                        start=start,
-                        end=end,
-                        text=(
-                            f"{one_text}\\N{two_text}"
-                            if one_text and two_text
-                            else one_text or two_text
-                        ),
-                    )
-                )
-            continue
-
-        # One to many mapping
-        if len(one_subs) == 1 and len(two_subs) > 1:
-            one_sub = one_subs[0]
-            timings = _get_sync_group_timings(
-                one_subs, two_subs, len(two_subs), timing_mode
-            )
-
-            for two_sub, (start, end) in zip(two_subs, timings):
-                one_text = one_sub.text
-                two_text = two_sub.text
-                synced.events.append(
-                    Subtitle(
-                        start=start,
-                        end=end,
-                        text=(
-                            f"{one_text}\\N{two_text}"
-                            if one_text and two_text
-                            else one_text or two_text
-                        ),
-                    )
-                )
-            continue
-
-        # Anything else is unsupported
-        raise ScinoephileError(
-            f"Unsupported sync group with {len(one_subs)} subtitles from series one "
-            f"and {len(two_subs)} subtitles from series two."
-        )
-
-    return _get_series_without_overlaps(synced)
-
-
-def _compare_sync_groups(  # noqa: PLR0912
-    first: SyncGroup, second: SyncGroup
-) -> int | None:
-    """Compare two sync groups.
-
-    Arguments:
-        first: First sync group
-        second: Second sync group
-    Returns:
-        -1 if first is less than second, 0 if they are equal, 1 if first is greater,
-        and None if they cannot be compared
-    """
-    first_min_one = min(first[0]) if first[0] else None
-    first_min_two = min(first[1]) if first[1] else None
-    second_min_one = min(second[0]) if second[0] else None
-    second_min_two = min(second[1]) if second[1] else None
-    first_order = None
-    second_order = None
-    if first_min_one is not None and second_min_one is not None:
-        if first_min_one < second_min_one:
-            first_order = -1
-        elif first_min_one > second_min_one:
-            first_order = 1
-        else:
-            first_order = 0
-    if first_min_two is not None and second_min_two is not None:
-        if first_min_two < second_min_two:
-            second_order = -1
-        elif first_min_two > second_min_two:
-            second_order = 1
-        else:
-            second_order = 0
-    match (first_order, second_order):
-        case (None, None):
-            return None
-        case (None, _):
-            return second_order
-        case (_, None):
-            return first_order
-        case (-1, -1):
-            return -1
-        case (0, 0):
-            return 0
-        case (1, 1):
-            return 1
-        case _:
-            raise ScinoephileError("Unexpected comparison result between sync groups")
-
-
-def _get_series_without_overlaps(series: Series) -> Series:
-    """Get a copy of a subtitle series with adjacent timing overlaps removed.
-
-    Arguments:
-        series: subtitle series whose events should be made non-overlapping
-    Returns:
-        Copy of the subtitle series with adjacent overlaps split at their midpoint
-    Raises:
-        ScinoephileError: if overlap removal would make a subtitle non-positive
-    """
-    output = type(series)(
-        events=[type(event)(**event.as_dict()) for event in series.events]
-    )
-
-    for idx, (previous, current) in enumerate(
-        zip(output.events, output.events[1:]), start=1
-    ):
-        if current.start >= previous.end:
-            continue
-
-        minimum_boundary = previous.start + 1
-        maximum_boundary = current.end - 1
-        if minimum_boundary > maximum_boundary:
-            raise ScinoephileError(
-                f"Cannot remove subtitle timing overlap between events {idx} "
-                f"and {idx + 1} without creating a non-positive duration "
-                f"subtitle: previous={previous.start}-{previous.end}, "
-                f"current={current.start}-{current.end}."
-            )
-
-        boundary = (previous.end + current.start) // 2
-        boundary = max(boundary, minimum_boundary)
-        boundary = min(boundary, maximum_boundary)
-
-        previous.end = boundary
-        current.start = boundary
-
-    return output
-
-
-def _get_sync_group_interval(
-    one_subs: list[Subtitle],
-    two_subs: list[Subtitle],
-    timing_mode: SyncTimingMode,
-) -> tuple[int, int]:
-    """Get the selected timing interval for a sync group.
-
-    Arguments:
-        one_subs: subtitles from the first series
-        two_subs: subtitles from the second series
-        timing_mode: timing mode used for synchronized subtitles
-    Returns:
-        Start and end times for the sync group
-    """
-    if timing_mode == SyncTimingMode.TOP and one_subs:
-        return one_subs[0].start, one_subs[-1].end
-    if timing_mode == SyncTimingMode.BOTTOM and two_subs:
-        return two_subs[0].start, two_subs[-1].end
-
-    starts = [sub.start for sub in one_subs + two_subs]
-    ends = [sub.end for sub in one_subs + two_subs]
-    return min(starts), max(ends)
-
-
-def _get_sync_group_timings(
-    one_subs: list[Subtitle],
-    two_subs: list[Subtitle],
-    count: int,
-    timing_mode: SyncTimingMode,
-) -> list[tuple[int, int]]:
-    """Get generated subtitle timings for a sync group.
-
-    Arguments:
-        one_subs: subtitles from the first series
-        two_subs: subtitles from the second series
-        count: number of generated subtitles
-        timing_mode: timing mode used for synchronized subtitles
-    Returns:
-        Start and end times for each generated subtitle
-    """
-    if timing_mode == SyncTimingMode.TOP and len(one_subs) == count:
-        return [(sub.start, sub.end) for sub in one_subs]
-    if timing_mode == SyncTimingMode.BOTTOM and len(two_subs) == count:
-        return [(sub.start, sub.end) for sub in two_subs]
-
-    # Split SyncTimingMode.OUTER, or top/bottom groups with mismatched output counts
-    group_start, group_end = _get_sync_group_interval(one_subs, two_subs, timing_mode)
-    edges = np.linspace(group_start, group_end, count + 1, dtype=int)
-    return [(int(start), int(end)) for start, end in zip(edges[:-1], edges[1:])]
-
-
-def _get_subtitle_group_midpoint(subtitles: list[Subtitle]) -> float:
-    """Get midpoint of a subtitle group span.
-
-    Arguments:
-        subtitles: subtitle group
-    Returns:
-        midpoint of the group span in milliseconds
-    """
-    return subtitles[0].start + (subtitles[-1].end - subtitles[0].start) / 2
-
-
-def _get_sync_groups(  # noqa: PLR0912, PLR0915
-    one: Series, two: Series, overlap: np.ndarray, cutoff: float
-) -> list[SyncGroup]:
-    """Build sync groups from a filtered overlap matrix.
-
-    Arguments:
-        one: first subtitle series
-        two: second subtitle series
-        overlap: overlap matrix between series
-        cutoff: overlap cutoff used when pruning matrix values
-    Returns:
-        list of sync groups derived from the overlap matrix
-    """
-    sync_groups = []
-
-    for i in range(len(one)):
-        scale = np.max(overlap[i])
-        for j in range(len(two)):
-            if overlap[i, j] / scale < cutoff:
-                overlap[i, j] = 0
-    for j in range(len(two)):
-        scale = np.max(overlap[:, j])
-        if scale > 0:
-            for i in range(len(one)):
-                if overlap[i, j] / scale < cutoff:
-                    overlap[i, j] = 0
-
-    logger.debug(f"OVERLAP ({cutoff:.2f}):\n{get_overlap_string(overlap)}")
-
-    available_is = set(range(len(one)))
-    available_js = set(range(len(two)))
-
-    nonzero = np.argwhere(overlap)
-
-    js_that_match_each_i = {}
-    for i in range(len(one)):
-        js_that_match_each_i[i] = [
-            int(j) for j in sorted(nonzero[nonzero[:, 0] == i, 1])
-        ]
-
-    is_that_match_each_j = {}
-    for j in range(len(two)):
-        is_that_match_each_j[j] = [
-            int(i) for i in sorted(nonzero[nonzero[:, 1] == j, 0])
-        ]
-
-    for i, js_that_match_this_i in js_that_match_each_i.items():
-        if i not in available_is:
-            continue
-
-        # One to zero
-        if len(js_that_match_this_i) == 0:
-            available_is.remove(i)
-            sync_groups.append(([i], []))
-            continue
-
-        if len(js_that_match_this_i) == 1:
-            j = js_that_match_this_i[0]
-            is_that_match_this_j = is_that_match_each_j[j]
-
-            # One to one
-            if len(is_that_match_this_j) == 1:
-                available_is.remove(i)
-                available_js.remove(j)
-                sync_groups.append(([i], [j]))
-                continue
-
-            # Many to one
-            if len(is_that_match_this_j) > 1:
-                if not all(i2 in available_is for i2 in is_that_match_this_j):
-                    raise ScinoephileError()
-                for i2 in is_that_match_this_j:
-                    available_is.remove(i2)
-                available_js.remove(j)
-                sync_groups.append((is_that_match_this_j, [j]))
-                continue
-
-        # One to many
-        if len(js_that_match_this_i) > 1:
-            for j in js_that_match_this_i:
-                is_that_match_this_j = is_that_match_each_j[j]
-                if len(is_that_match_this_j) != 1:
-                    raise ScinoephileError()
-            available_is.remove(i)
-            for j in js_that_match_this_i:
-                available_js.remove(j)
-            sync_groups.append(([i], js_that_match_this_i))
-            continue
-
-    # Raise exception if there are any remaining subtitles in series one
-    if len(available_is) > 0:
-        raise ScinoephileError()
-
-    # Add remaining subtitles from series two
-    for j in available_js:
-        sync_groups.extend([([], [j])])
-
-    # Sort sync groups by their indexes
-    sync_groups = _sort_sync_groups(sync_groups)
-
-    logger.info(f"OVERLAP ({cutoff:.2f}):\n{get_overlap_string(overlap)}")
-
-    return sync_groups
-
-
-def _sort_sync_groups(sync_groups: list[SyncGroup]) -> list[SyncGroup]:
-    """Sort sync groups.
-
-    May not correctly handle all initial orders, but should work for the cases
-    encountered.
-
-    Arguments:
-        sync_groups: Sync groups to sort
-    Returns:
-        Sorted sync groups
-    """
-    sorted_groups = []
-
-    for group in sync_groups:
-        inserted = False
-        for i in range(len(sorted_groups) + 1):
-            # Try inserting at position i
-            if i == len(sorted_groups):
-                sorted_groups.append(group)
-                inserted = True
-                break
-
-            result = _compare_sync_groups(group, sorted_groups[i])
-            if result is None:
-                continue  # Try comparing with the next one
-            elif result < 0:
-                sorted_groups.insert(i, group)
-                inserted = True
-                break
-        if not inserted:
-            raise ScinoephileError(
-                "Could not determine correct position for sync group"
-            )
-
-    return sorted_groups
